+++ src/Psl/Async/Exception/UnhandledAwaitableException.php
@@ -4,18 +4,18 @@ declare(strict_types=1);
 
 namespace Psl\Async\Exception;
 
-use Exception as RootException;
 use Psl\Exception\RuntimeException;
 use Psl\Str;
+use Throwable;
 
 final class UnhandledAwaitableException extends RuntimeException implements ExceptionInterface
 {
-    public static function forException(RootException $exception): UnhandledAwaitableException
+    public static function forThrowable(Throwable $throwable): UnhandledAwaitableException
     {
         return new self(
-            Str\format('Unhandled awaitable error "%s", make sure to call `Awaitable::await()` before the awaitable is destroyed, or call `Awaitable::ignore()` to ignore exceptions.', $exception::class),
-            (int) $exception->getCode(),
-            $exception
+            Str\format('Unhandled awaitable error "%s", make sure to call `Awaitable::await()` before the awaitable is destroyed, or call `Awaitable::ignore()` to ignore exceptions.', $throwable::class),
+            (int) $throwable->getCode(),
+            $throwable
         );
     }
 }